UserService,TaskService

+++ progettofinale/progettofinale/src/main/java/it/itconsulting/progettofinale/repository/AppUserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.itconsulting.progettofinale.model.AppUser;
 
-public interface AppUserRepository extends JpaRepository<AppUser, Integer>{
+public interface AppUserRepository extends JpaRepository<AppUser, Long>{
     
 }
